Add files via upload
Evolution is the 'hub' token of our project ecosystem. Evolution token will allow us to provide staking functions for CardanoEvo (and additional tokens in the future), encourage further investment and deliver NFT minting & market-placing. The Evolution ‘hub’ token will also provide future functionality to symbiotic tokens within our own ecosystem, by allowing staking and liquidity provision. Lastly, the Evolution ‘hub’ token will provide direct liquidity to CardanoEvo and our future symbiont tokens, thus providing long-term stability.

It may seem complicated, because it certainly is! We are in fact building an ecosystem which no one else has, or is in the process of attempting; one specifically designed to be to expand, evolve and scale. A lot of hard work has gone into this project to deliver the next step in the Evolution ecosystem.

We are Evolution.
We are CardanoEvo.
We are Community, Evolved.